+++ src/components/pages/LoginPage/LoginForm/index.js
@@ -1,0 +1,1 @@
+export * from "./LoginForm";
