+++ src/components/pages/LoginPage/LoginForm/style.jsx
@@ -1,0 +1,23 @@
+import { styled } from "styled-components";
+
+export const FormBox = styled.form`
+  width: 300px;
+`;
+
+export const SNSLoginBox = styled.div`
+  width: 100%;
+`;
+
+export const SNSLoginText = styled.span`
+  margin: 35px 0 15px;
+  color: #757575;
+  font-size: 12px;
+  font-weight: normal;
+  line-height: 1.33;
+  text-align: center;
+`;
+
+export const SNSLoginBtn = styled.span`
+  width: 100%;
+  cursor: pointer;
+`;
