+++ src/components/pages/index.js
@@ -1,1 +1,2 @@
 export * from "./ItemDetailPage";
+export * from "./LoginPage";
