+++ src/pages/LoginPage/style.jsx
@@ -28,10 +28,6 @@ export const LogoBox = styled.div`
   height: 52.5px;
 `;
 
-export const FormBox = styled.form`
-  width: 300px;
-`;
-
 export const UserAccountForm = styled.div`
   width: 100%;
   text-align: center;
@@ -41,24 +37,6 @@ export const UserAccountForm = styled.div`
 `;
 
 export const UserAccountBtn = styled.span`
-  width: 100%;
-  cursor: pointer;
-`;
-
-export const SNSLoginBox = styled.div`
-  width: 100%;
-`;
-
-export const SNSLoginText = styled.span`
-  margin: 35px 0 15px;
-  color: #757575;
-  font-size: 12px;
-  font-weight: normal;
-  line-height: 1.33;
-  text-align: center;
-`;
-
-export const SNSLoginBtn = styled.span`
   width: 100%;
   cursor: pointer;
 `;
